chore(package): replace jest with mocha + chai
Installing jest was a mistake. I faced some difficulties that has already been
resolved in mocha. As an example, good support for async functions or promises.
The only benefit in jest was auto-mocking dependencies but there are a few real
bottlenecks (fs, net, and DB) in Umzug. I think, it is possible to isolate them
out in another way.